new sharding approach

+++ finetuning/train1.py
@@ -454,7 +454,7 @@ def _train_step(state, batch, rng, model):
     def _loss(p):
 
         batch_size, _ = batch["inputs"].shape
-        cache = model.init_cache(batch_size, dtype=model.config.dtype)
+        cache = model.init_cache(batch_size, dtype=config.dtype)
         logits = state.apply_fn(
             {"params": p},
             tokens=batch["inputs"],
@@ -474,7 +474,7 @@ def _train_step(state, batch, rng, model):
 def _eval_step(state, batch, model):
     def _loss(p):
         batch_size, _ = batch["inputs"].shape
-        cache = model.init_cache(batch_size, dtype=model.config.dtype)
+        cache = model.init_cache(batch_size, dtype=config.dtype)
         logits = state.apply_fn(
             {"params": p},
             tokens=batch["inputs"],
